something is missing.
It works. but it can not handle the values in order to move on the next
stage. It is somethign about the variable logic or the str type. Maybe i
will start continue with the generator, and i will fix this problem
later, maybe i will just start commenting every part of this code.